Extract Wallet interface into separate file
