+++ packages/iov-keycontrol/src/wallets/ed25519hdwallet.ts
@@ -1,6 +1,6 @@
 import { Slip10Curve } from "@iov/crypto";
 
-import { WalletImplementationIdString } from "../keyring";
+import { WalletImplementationIdString } from "../wallet";
 import { Slip10Wallet } from "./slip10wallet";
 
 export class Ed25519HdWallet extends Slip10Wallet {
