+++ packages/iov-keycontrol/types/keyring.d.ts
@@ -1,21 +1,6 @@
 import { As } from "type-tagger";
-import { PrehashType, SignableBytes } from "@iov/bcp-types";
-import { Slip10RawIndex } from "@iov/crypto";
-import { ValueAndUpdates } from "@iov/stream";
-import { ChainId, PublicKeyBundle, SignatureBytes } from "@iov/tendermint-types";
-import { Ed25519Wallet } from "./wallets";
+import { Wallet, WalletId, WalletImplementationIdString, WalletSerializationString } from "./wallet";
 export declare type KeyringSerializationString = string & As<"keyring-serialization">;
-export declare type WalletImplementationIdString = string & As<"wallet-implementation-id">;
-export declare type WalletSerializationString = string & As<"wallet-serialization">;
-export declare type LocalIdentityId = string & As<"local-identity-id">;
-export declare type WalletId = string & As<"wallet-id">;
-export interface PublicIdentity {
-    readonly pubkey: PublicKeyBundle;
-}
-export interface LocalIdentity extends PublicIdentity {
-    readonly id: LocalIdentityId;
-    readonly label?: string;
-}
 export declare type WalletDeserializer = (data: WalletSerializationString) => Wallet;
 /**
  * A collection of wallets
@@ -41,25 +26,4 @@ export declare class Keyring {
     getWallet(id: WalletId): Wallet | undefined;
     serialize(): KeyringSerializationString;
     clone(): Keyring;
-}
-/**
- * A is a generic interface for managing a set of keys and signing
- * data with them.
- *
- * A Wallet is responsible for storing private keys securely and
- * signing with them. Wallet can be implemented in software or as
- * a bridge to a hardware wallet.
- */
-export interface Wallet {
-    readonly label: ValueAndUpdates<string | undefined>;
-    readonly id: WalletId;
-    readonly setLabel: (label: string | undefined) => void;
-    readonly createIdentity: (options: Ed25519Wallet | ReadonlyArray<Slip10RawIndex> | number) => Promise<LocalIdentity>;
-    readonly setIdentityLabel: (identity: PublicIdentity, label: string | undefined) => void;
-    readonly getIdentities: () => ReadonlyArray<LocalIdentity>;
-    readonly canSign: ValueAndUpdates<boolean>;
-    readonly implementationId: WalletImplementationIdString;
-    readonly createTransactionSignature: (identity: PublicIdentity, transactionBytes: SignableBytes, prehash: PrehashType, chainId: ChainId) => Promise<SignatureBytes>;
-    readonly serialize: () => WalletSerializationString;
-    readonly clone: () => Wallet;
 }
