+++ packages/iov-keycontrol/types/userprofile.d.ts
@@ -4,7 +4,8 @@ import { ReadonlyDate } from "readonly-date";
 import { Nonce, SignedTransaction, TxCodec, UnsignedTransaction } from "@iov/bcp-types";
 import { Slip10RawIndex } from "@iov/crypto";
 import { ValueAndUpdates } from "@iov/stream";
-import { Keyring, LocalIdentity, PublicIdentity, Wallet, WalletId } from "./keyring";
+import { Keyring } from "./keyring";
+import { LocalIdentity, PublicIdentity, Wallet, WalletId } from "./wallet";
 import { Ed25519Wallet } from "./wallets";
 export interface UserProfileOptions {
     readonly createdAt: ReadonlyDate;
