+++ packages/iov-keycontrol/types/wallets/slip10wallet.d.ts
@@ -2,7 +2,7 @@ import { PrehashType, SignableBytes } from "@iov/bcp-types";
 import { Slip10Curve } from "@iov/crypto";
 import { ValueAndUpdates } from "@iov/stream";
 import { ChainId, SignatureBytes } from "@iov/tendermint-types";
-import { LocalIdentity, PublicIdentity, Wallet, WalletId, WalletImplementationIdString, WalletSerializationString } from "../keyring";
+import { LocalIdentity, PublicIdentity, Wallet, WalletId, WalletImplementationIdString, WalletSerializationString } from "../wallet";
 interface Slip10WalletConstructor {
     new (data: WalletSerializationString): Slip10Wallet;
 }
